WIP: fix structural issues for good
