--- components/staff_schedule/IMPLEMENTATION.log
@@ -1,59 +1,0 @@
-2026-06-19
-- Reviewed current staff_schedule frontend API draft against legacy work_schedule.
-- Confirmed the draft is a compatibility contract, not yet a final frontend-facing schema.
-- Saved concrete cleanup/add/drop recommendations in API_SUGGESTIONS.md.
-- Next planned step: start desktop UI shell on isolated route with mock-first data source and adapter boundary for the new API.
-- Switched plan: no mock stage, live staff_schedule API only.
-- Trimmed API_SUGGESTIONS.md down to first-render BE needs only.
-- Added the first desktop slice: route, live API wrapper, page hook, and read-only schedule table shell wired to get_all/get_graph.
-
-- Switched plan: no mock stage, live staff_schedule API only.
-- Next: verify authenticated payloads from localhost and start desktop route/page implementation.
-- Found runtime blocker: useApi hits http://localhost:8000/api/staff_schedule/get_all and gets 403; app redirects to /.
-- Next: switch staff_schedule wrapper to 127.0.0.1 transport used by legacy pages so desktop render can proceed.
-- Switched the module to a local non-redirecting transport so the route stays mounted on API errors.
-- Verified `/staff_schedule` renders the desktop shell, filters, period tabs, and error/empty states.
-- Current blocker: `POST http://127.0.0.1:8000/api/staff_schedule/get_all` returns `403 Forbidden` for the authenticated localhost session.
-- Verified new `staff_schedule/get_all` is live, but backend wraps the payload under top-level `data`; adjusted FE transport to unwrap it.
-- Verified `staff_schedule/get_graph` still returns `500`, so read-only graph/detail requests now fall back in isolation to legacy `work_schedule`.
-- Desktop table now renders real data with point/month filters, period tabs, summary columns, and footer metrics.
-- Fixed duplicate React row keys and corrected footer mapping for rolls/pizza/order metrics.
-- Fixed local staff_schedule transport unwrap: backend returns response under top-level data.
-2026-06-22 10:04 +0300
-- reviewed staff_schedule against ads/vendors/cleanings module style
-- switched primary API path to shared useApi("staff_schedule"), kept isolated legacy fallback only for broken read methods
-- removed temporary window debug globals from page state/api loading flow
-- cleaned minor page drift: memoized handlers remain, unused ScheduleRow prop removed
-2026-06-22 10:31 +0300
-- switched hanging graph reads to timeout-based legacy fallback on shared useApi path
-- flattened main screen shell toward legacy/Figma base: removed extra subtitle framing, tightened title/tabs/filter layout
-- tightened main table proportions toward legacy base (sticky cols, day cells, summary cols, row spacing)
-- removed visible source/debug meta from main toolbar area
-- normalized fallback module title to keep Graph title stable during legacy bootstrap
-- removed full-page graph-loading blocker; table area now owns the loading state
-- renamed bootstrap contract/docs to `months`
-- renamed graph request field/docs to `month`
-- fixed `hours_days` naming in local staff_schedule contract docs
-2026-06-23
-- updated PLAN.md with the modal track: Figma/legacy modal rules, shared component usage, reusable mobile `MyDrawer` wrapper, and staged day/month/smena modal plan
-- removed the non-Figma stats/meta strip from the main staff_schedule header
-- added shared mobile `MyDrawer` and responsive modal scaffolding for staff_schedule
-- wired first read-only modal slice: day modal from day cells and month modal from employee name
-- removed the explicit mobile drawer close button; mobile modal dismissal is now swipe/backdrop-driven
-- switched the month modal toward legacy/Figma behavior by reusing the shared calendar picker instead of a custom-only list view
-- synced the local staff_schedule API wrapper to the expanded FE contract method set from `fe-api.md`
-- re-reviewed legacy `work_schedule` modal flow and confirmed the current read-only day/month modals are insufficient
-- updated PLAN.md to replace the read-only modal assumption with an edit-first staged modal track based on legacy behavior and the current Figma direction
-- rebuilt the day modal into an editable shell with shared controls, optional app/mentor fields, health block, hours editor, history block, and responsive save/cancel footer
-- wired day modal submit to live `save_user_day` with graph refresh on success and kept all changes isolated to `staff_schedule`
-- reviewed the provided desktop screen and formalized the purpose of the missing control layer before coding it
-- added the desktop control shell: top filters, action icons shell, period switch, shift filter, journal button shell, table toolbar, row checkboxes, and per-shift collapse
-- wired local desktop-only behavior for shift filtering, calendar visibility, color-mode presentation, and row selection without expanding backend scope
-- refactored the desktop header from rigid grid fractions to flexible action-aligned rows so the refresh cluster matches the base shell
-- removed the double-outline wrapper around the color legend selector and corrected the calendar toggle control shell
-- replaced the desktop modal wrapper in `staff_schedule` with a module-local dialog header so long titles keep the close action pinned on the right
-2026-06-22 18:00 +0300
-- authenticated Figma MCP and located the target design page `График работы`, but further design inspection is blocked by the current Figma MCP seat rate limit
-- adjusted near-term execution plan: keep the module read-only and UI-focused, avoid inventing mutation behavior until backend routes land
-- refactored `staff_schedule` page structure toward project module patterns: extracted header/table sections, added schedule constants, and moved graph/page derivation into a dedicated view-model helper
-- kept current rendering behavior and live/fallback read flow intact so future desktop/mobile parity work can build on a cleaner module shape
