--- components/staff_schedule/TASK.md
@@ -1,18 +1,0 @@
-https://www.figma.com/design/bzb8ksK03R5FeMzRBKsEo9/Chief-screens?node-id=380-206795&t=wWGzyWi8mJvb7HgR-0
-
-https://jacosoft-dop.ru/work_schedule - оригинальный модуль
-
-создаем новый модуль графика работы, и работаем пока в нем
-таблицы тоже новые создаем, можно сразу с данными
-cafe_smena_info - информация о 15 днях работы ( зп, бонусы, штрафы и прочее )
-cafe_smena_days - информация по дневным сменам ( нагрузка, часы )
-cafe_smena_hours - именно часы работы
-
-- таблицы с историей
-  cafe_smena - все смены с привязкой к кафе
-  cafe_smena_auth - информации об авторизации, скорей всего с кухни заполняется
-  cafe_smena_users - сотрудники в смене
-
-пока переносим чисто отображение данных, без обработок, продления графика, расстановки часов и прочего
-
-в новых таблицах надо расставить ключи так, чтобы смены не дублировались по умолчанию ( сейчас это главный косяк текущей реализации, она сейчас частично решена проверками, но ключами будет проще )
