+++ ui/v2/Autocomplete.jsx
@@ -1,9 +1,11 @@
 import { MyAutocomplite2 } from "@/ui/Forms";
 
-export default function V2Autocomplete(props) {
+export default function V2Autocomplete({ options, data, onChange, func, ...props }) {
   return (
     <MyAutocomplite2
       unifiedPopup
+      data={options ?? data}
+      func={onChange ?? func}
       {...props}
     />
   );
