+++ ui/v2/DatePicker.jsx
@@ -1,5 +1,10 @@
 import { MyDatePickerNew } from "@/ui/Forms";
 
-export default function V2DatePicker(props) {
-  return <MyDatePickerNew {...props} />;
+export default function V2DatePicker({ onChange, func, ...props }) {
+  return (
+    <MyDatePickerNew
+      func={onChange ?? func}
+      {...props}
+    />
+  );
 }
