+++ ui/v2/TextInput.jsx
@@ -1,5 +1,10 @@
 import { MyTextInput } from "@/ui/Forms";
 
-export default function V2TextInput(props) {
-  return <MyTextInput {...props} />;
+export default function V2TextInput({ onChange, func, ...props }) {
+  return (
+    <MyTextInput
+      func={onChange ?? func}
+      {...props}
+    />
+  );
 }
