+++ ui/v2/TimePicker.jsx
@@ -1,5 +1,10 @@
 import { MyTimePicker } from "@/ui/Forms";
 
-export default function V2TimePicker(props) {
-  return <MyTimePicker {...props} />;
+export default function V2TimePicker({ onChange, func, ...props }) {
+  return (
+    <MyTimePicker
+      func={onChange ?? func}
+      {...props}
+    />
+  );
 }
